requirements.txt
Django>=5.2,<5.3
Pillow>=10.4,<12.0
open-clip-torch>=2.32,<3.0
torch>=2.6,<2.8
transformers>=4.51,<5.0
sentencepiece>=0.2.0,<1.0
